+++ test/helper.go
@@ -44,160 +44,12 @@ var Events = []models.Event{
 	},
 }
 
-type DbConfig struct {
-	Client *sql.DB
-}
-
-type SQSReceiveMessageImpl struct{}
-
-func (dt SQSReceiveMessageImpl) GetQueueUrl(ctx context.Context,
-	params *sqs.GetQueueUrlInput,
-	optFns ...func(*sqs.Options)) (*sqs.GetQueueUrlOutput, error) {
-
-	output := &sqs.GetQueueUrlOutput{
-		QueueUrl: aws.String(app.QueueName),
-	}
-
-	return output, nil
-}
-
-func (dt SQSReceiveMessageImpl) ReceiveMessage(ctx context.Context,
-	params *sqs.ReceiveMessageInput,
-	optFns ...func(*sqs.Options)) (*sqs.ReceiveMessageOutput, error) {
-
-	messageOne := Events[0]
-	messageOneMarshalled, _ := json.Marshal(messageOne)
-
-	messageTwo := Events[1]
-	messageTwoMarshalled, _ := json.Marshal(messageTwo)
-
-	messages := []types.Message{
-		{
-			MessageId:     aws.String("message-one-id"),
-			ReceiptHandle: aws.String("message-one-receipt-handle"),
-			Body:          aws.String(string(messageOneMarshalled)),
-		},
-		{
-			MessageId:     aws.String("message-two-id"),
-			ReceiptHandle: aws.String("message-two-receipt-handle"),
-			Body:          aws.String(string(messageTwoMarshalled)),
-		},
-	}
-
-	output := &sqs.ReceiveMessageOutput{
-		Messages: messages,
-	}
-
-	return output, nil
-}
-
 func createHeaders() map[string]string {
 	headers := make(map[string]string)
 
 	headers["Accept"] = "application/json"
 
 	return headers
-}
-
-func DbClient() *sql.DB {
-	dbDsn := os.Getenv("EVENTS_DB_DSN")
-
-	logger := log.New(os.Stdout, "", log.Ldate|log.Ltime)
-
-	db, err := sql.Open("postgres", dbDsn)
-	if err != nil {
-		logger.Fatal(err)
-	}
-
-	err = db.Ping()
-	if err != nil {
-		logger.Fatal(err)
-	}
-
-	logger.Printf("database connection pool established")
-
-	return db
-}
-
-func (config *DbConfig) CreateDbTable() error {
-	query := `
-		CREATE TABLE IF NOT EXISTS events(
-			id uuid PRIMARY KEY UNIQUE DEFAULT gen_random_uuid() NOT NULL,
-			device_id VARCHAR NOT NULL,
-			description VARCHAR NOT NULL,
-			type VARCHAR NOT NULL ,
-			event VARCHAR NOT NULL,
-			read BOOLEAN DEFAULT FALSE,
-			created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-			updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-		);
-		
-		CREATE INDEX updated_at ON events (updated_at);
-	`
-
-	_, err := config.Client.Query(query)
-	if err != nil {
-		fmt.Errorf("DeleteEventsErr %v", err)
-		return err
-	}
-
-	return nil
-}
-
-func (config *DbConfig) DropDbTable() error {
-	query := `
-		DROP TABLE IF EXISTS events
-	`
-
-	_, err := config.Client.Query(query)
-	if err != nil {
-		fmt.Errorf("DeleteEventsErr %v", err)
-		return err
-	}
-
-	return nil
-}
-
-func (config *DbConfig) DeleteEvents() error {
-	query := `
-		DELETE FROM events
-	`
-
-	_, err := config.Client.Query(query)
-	if err != nil {
-		fmt.Errorf("DeleteEventsErr %v", err)
-		return err
-	}
-
-	return nil
-}
-
-func (config *DbConfig) CreateEvent(data models.Event) (models.Event, error) {
-	query := `
-		INSERT INTO events (device_id, description, type, event, read)
-		VALUES ($1, $2, $3, $4, $5)
-		RETURNING id, device_id AS deviceId, description, type, event, read
-	`
-
-	var event models.Event
-
-	args := []interface{}{data.DeviceId, data.Description, data.Type, data.Event, data.Read}
-	row := config.Client.QueryRow(query, args...)
-
-	err := row.Scan(
-		&event.Id,
-		&event.DeviceId,
-		&event.Description,
-		&event.Type,
-		&event.Event,
-		&event.Read,
-	)
-
-	if err != nil {
-		return models.Event{}, err
-	}
-
-	return event, err
 }
 
 func StartServer() *http.Server {
@@ -231,6 +83,7 @@ func StartServer() *http.Server {
 		MessageAttributeNames: []string{
 			string(types.QueueAttributeNameAll),
 		},
+
 		QueueUrl:            queueURL,
 		MaxNumberOfMessages: 25,
 		VisibilityTimeout:   int32(timeout),
